updated file structure for feed pages - all feed components now live directly under the Feed component folder
