--- client/src/components/Feed/delayFeed/index.js
@@ -1,1 +1,0 @@
-export { default } from "./delayFeed.js";
