+++ client/src/components/Feed/index.js
@@ -1,1 +1,6 @@
 export { default } from "./Feed.js";
+export * from "./AllFeed.js";
+export * from "./delayFeed.js";
+export * from "./trainFeed.js";
+export * from "./stationFeed.js";
+export * from "./myFeed.js";
