--- client/src/components/Feed/myFeed/index.js
@@ -1,1 +1,0 @@
-export { default } from "./myFeed.js";
