--- client/src/components/Feed/stationFeed/index.js
@@ -1,1 +1,0 @@
-export { default } from "./stationFeed.js";
