--- client/src/components/Feed/trainFeed/index.js
@@ -1,1 +1,0 @@
-export { default } from "./trainFeed.js";
